Refactored code editor from collaboration page
