+++ frontend/pages/collaboration.tsx
@@ -1,16 +1,10 @@
 "use client";
-import { Box, Flex, HStack, Spacer, Text } from "@chakra-ui/react";
-import { useState, useRef, useEffect } from "react";
-import Editor, { DiffEditor, useMonaco, loader } from "@monaco-editor/react";
-import { editor } from "monaco-editor";
-import { io } from "socket.io-client";
-import { Select } from "@chakra-ui/react";
+import { Box, Text } from "@chakra-ui/react";
+import { useState } from "react";
+import CodeEditor from "../components/CodeEditor";
 import VideoCall from "../components/VideoCall";
 
 export default function Collaboration() {
-  const editorRef = useRef();
-  const [socket, setSocket] = useState(null);
-  const isIncomingCode = useRef(false);
   const [questions, setQuestions] = useState([]);
   const selectedItem = {
     name: "Two Sum",
@@ -19,59 +13,9 @@ export default function Collaboration() {
     createdAt: "2022-10-22 20:19:50.236312+00:00",
     updatedAt: "2022-10-22 20:19:50.236312+00:00",
   };
-  const handleEditorDidMount = (editor: editor.IStandaloneCodeEditor) => {
-    editorRef.current = editor;
-  };
-  const selectRef = useRef(null);
-  const handleLanguageChange = (e) => {
-    window.monaco.editor.setModelLanguage(
-      editorRef.current?.getModel(),
-      e.target.value
-    );
-    socket?.emit("languageChange", e.target.value);
-  };
-
-  const handleChange = (
-    value = "",
-    event: editor.IModelContentChangedEvent
-  ) => {
-    if (isIncomingCode.current) {
-      isIncomingCode.current = false;
-      return;
-    }
-    console.log("sent");
-    socket?.emit("codeChange", event);
-  };
-
-  useEffect(() => {
-    const socket = io("http://localhost:8080");
-    setSocket(socket);
-    socket.on("codeChange", (event) => {
-      isIncomingCode.current = true;
-      console.log("received", event);
-      editorRef.current?.getModel()?.applyEdits(event.changes);
-    });
-    socket.on("languageChange", (event) => {
-      console.log("received", event);
-      window.monaco.editor.setModelLanguage(
-        editorRef.current?.getModel(),
-        event
-      );
-      selectRef.current.value = event;
-    });
-    return () => {
-      socket.disconnect();
-    };
-  }, []);
 
   return (
     <div>
-      <select ref={selectRef} onChange={handleLanguageChange}>
-        <option value="javascript">JS</option>
-        <option value="python">Python</option>
-        <option value="C++">C++</option>
-        <option value="Java">Java</option>
-      </select>
       <Box display="flex" height="100vh">
         <Box flexDirection="column">
           {selectedItem && (
@@ -82,15 +26,7 @@ export default function Collaboration() {
           <VideoCall />
         </Box>
         <Box flex={1} m="30px">
-          <Editor
-            onChange={handleChange}
-            height="100vh"
-            width="50vw"
-            theme="vc"
-            onMount={handleEditorDidMount}
-            defaultLanguage="javascript"
-            defaultValue="// some comment"
-          />
+          <CodeEditor />
         </Box>
       </Box>
     </div>
